회원가입 조건 추가 , 토큰 검사 후 게시글 작성/수정/삭제 구현

+++ src/main/java/com/sparta/blog/service/BlogService.java
@@ -4,24 +4,32 @@ import com.sparta.blog.dto.BlogRequestDto;
 import com.sparta.blog.dto.BlogResponseDto;
 import com.sparta.blog.dto.StatusCodeDto;
 import com.sparta.blog.entity.Blog;
+import com.sparta.blog.jwt.JwtUtil;
 import com.sparta.blog.repository.BlogRepository;
+import io.jsonwebtoken.Claims;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.CookieValue;
 
 import java.util.List;
 
 @Service
 public class BlogService {
     private final BlogRepository blogRepository;
+    private final JwtUtil jwtUtil;
 
-    public BlogService(BlogRepository blogRepository){
+    public BlogService(BlogRepository blogRepository, JwtUtil jwtUtil){
         this.blogRepository = blogRepository;
+        this.jwtUtil = jwtUtil;
     }
 
-    public BlogResponseDto createBlog(BlogRequestDto requestDto) {
+    public BlogResponseDto createBlog(BlogRequestDto requestDto, @CookieValue(JwtUtil.AUTHORIZATION_HEADER) String tokenValue) {
+        //token에서 loginname 가져오기
+        String loginname = loginUsername(tokenValue);
+
         //RequestDto -> Entity
-        Blog blog = new Blog(requestDto);
+        Blog blog = new Blog(requestDto, loginname);
 
         //DB 저장
         Blog saveBlog = blogRepository.save(blog);
@@ -38,23 +46,34 @@ public class BlogService {
     }
 
     @Transactional
-    public BlogResponseDto updateBlog(Long id, BlogRequestDto requestDto) {
+    public BlogResponseDto updateBlog(Long id, BlogRequestDto requestDto, @CookieValue(JwtUtil.AUTHORIZATION_HEADER) String tokenValue) {
         //해당 글이 DB에 존재하는지 확인
         Blog blog = findBlog(id);
-        //비밀번호 확인
-        blog = passCheck(blog, requestDto);
 
+        //token에서 loginname 가져오기
+        String loginname = loginUsername(tokenValue);
+
+        // username 비교
+        if (!loginname.equals(blog.getUsername())) {
+            throw new IllegalArgumentException("작성자가 아닙니다.");
+        }
         //글 내용 수정
         blog.update(requestDto);
 
         return new BlogResponseDto(blog);
     }
 
-    public StatusCodeDto deleteBlog(Long id, BlogRequestDto requestDto) {
+    public StatusCodeDto deleteBlog(Long id, @CookieValue(JwtUtil.AUTHORIZATION_HEADER) String tokenValue) {
         //해당 글이 DB에 존재하는지 확인
         Blog blog = findBlog(id);
-        //비밀번호 확인
-        blog = passCheck(blog, requestDto);
+
+        //token에서 loginname 가져오기
+        String loginname = loginUsername(tokenValue);
+
+        // username 비교
+        if (!loginname.equals(blog.getUsername())) {
+            throw new IllegalArgumentException("작성자가 아닙니다.");
+        }
 
         //해당 글 삭제하기
         blogRepository.delete(blog);
@@ -79,11 +98,18 @@ public class BlogService {
         return blog;
     }
 
-    private Blog passCheck(Blog blog, BlogRequestDto requestDto){
-        //비밀번호 확인
-        if(blog.getPassword() != null && !blog.getPassword().equals(requestDto.getPassword())){
-            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+    private String loginUsername(@CookieValue(JwtUtil.AUTHORIZATION_HEADER) String tokenValue) {
+        // JWT 토큰 substring
+        String token = jwtUtil.substringToken(tokenValue);
+        // 토큰 검증
+        if(!jwtUtil.validateToken(token)){
+            throw new IllegalArgumentException("Token Error");
         }
-        return blog;
+        // 토큰에서 사용자 정보 가져오기
+        Claims info = jwtUtil.getUserInfoFromToken(token);
+        // 사용자 username
+        String loginname = info.getSubject();
+
+        return loginname;
     }
 }
